add update user by id and refactor code

+++ server/controllers/user.js
@@ -1,4 +1,3 @@
-import bcrypt from "bcrypt";
 import { v2 as cloudinary } from "cloudinary";
 import { avatarOptions } from "../constants.js";
 import User from "../models/User.js";
@@ -6,7 +5,7 @@ import hashPassword from "../utils/hashPassword.js";
 import sendError from "../utils/sendError.js";
 import sendSuccess from "../utils/sendSuccess.js";
 
-export const getAllUsers = async (req, res, next) => {
+export const getAll = async (req, res, next) => {
 	try {
 		// Get all users
 		const users = await User.find().select("-__v -password");
@@ -19,7 +18,7 @@ export const getAllUsers = async (req, res, next) => {
 	}
 };
 
-export const getUserById = async (req, res, next) => {
+export const getById = async (req, res, next) => {
 	// Get user id from request params
 	const { id } = req.params;
 
@@ -35,7 +34,7 @@ export const getUserById = async (req, res, next) => {
 	}
 };
 
-export const getUserProfile = async (req, res, next) => {
+export const getProfile = async (req, res, next) => {
 	// Get user id from request
 	const { userId } = req;
 
@@ -51,8 +50,8 @@ export const getUserProfile = async (req, res, next) => {
 	}
 };
 
-export const updateUserProfile = async (req, res, next) => {
-	// Get user if from request
+export const updateProfile = async (req, res, next) => {
+	// Get user id from request
 	const { userId } = req;
 	// Get data from request body
 	const { email, currentPassword, phone } = req.body;
@@ -62,40 +61,6 @@ export const updateUserProfile = async (req, res, next) => {
 	try {
 		// Get user by id
 		const user = await User.findById(userId);
-		if (!user) return sendError(res, "User not found", 404);
-
-		// Compare password with current password
-		const comparedPassword = bcrypt.compareSync(currentPassword, user.password);
-		if (!comparedPassword)
-			return sendError(
-				res,
-				"Sorry, the password you entered is incorrect. Please try again",
-				400,
-				"currentPassword"
-			);
-
-		// Check email exists or not in database
-		const isEmailExists = await User.findOne({
-			email: { $eq: email, $ne: user.email },
-		});
-		if (isEmailExists)
-			return sendError(
-				res,
-				`User with this email (${isEmailExists.email}) already exists`,
-				409,
-				"email"
-			);
-		// Check phone exists or not in database
-		const isPhoneExists = await User.findOne({
-			phone: { $eq: phone, $ne: user.phone },
-		});
-		if (isPhoneExists)
-			return sendError(
-				res,
-				`User with this phone number (${isPhoneExists.phone}) already exists`,
-				409,
-				"phone"
-			);
 
 		// Check file exist or not
 		if (file) {
@@ -123,7 +88,49 @@ export const updateUserProfile = async (req, res, next) => {
 		}
 
 		// Send success notification
-		return sendSuccess(res, "Successfully edited personal profile");
+		return sendSuccess(res, "Your profile has been edited successfully");
+	} catch (error) {
+		next(error);
+	}
+};
+
+export const updateById = async (req, res, next) => {
+	// Get user id from request
+	const { id } = req.params;
+	// Get file from request
+	const { file } = req;
+
+	try {
+		// Get user by id
+		const user = await User.findById(id);
+
+		// Check file exist or not
+		if (file) {
+			const { path } = file;
+
+			await cloudinary.uploader
+				.upload(path, avatarOptions)
+				.then(async (result) => {
+					await User.findByIdAndUpdate(
+						id,
+						{
+							...req.body,
+							password: user.password,
+							avatar: result.secure_url,
+						},
+						{ new: true }
+					);
+				});
+		} else {
+			await User.findByIdAndUpdate(
+				id,
+				{ ...req.body, password: user.password },
+				{ new: true }
+			);
+		}
+
+		// Send success notification
+		return sendSuccess(res, "Edited user successfully");
 	} catch (error) {
 		next(error);
 	}
@@ -133,37 +140,9 @@ export const changePassword = async (req, res, next) => {
 	// Get user id from request
 	const { userId } = req;
 	// Get data from request body
-	const { currentPassword, newPassword } = req.body;
+	const { newPassword } = req.body;
 
 	try {
-		// Get user by id
-		const user = await User.findById(userId);
-		if (!user) return sendError(res, "User not found", 404);
-
-		// Compare current password
-		const isMatchCurrentPassword = bcrypt.compareSync(
-			currentPassword,
-			user.password
-		);
-		// Check if the password entered by the user does not match the password in the database
-		if (!isMatchCurrentPassword)
-			return sendError(
-				res,
-				"Sorry, the current password you entered is incorrect. Please try again",
-				400,
-				"currentPassword"
-			);
-
-		// Compare new password
-		const isMatchNewPassword = bcrypt.compareSync(newPassword, user.password);
-		if (isMatchNewPassword)
-			return sendError(
-				res,
-				"The new password must be different from the old password",
-				400,
-				"newPassword"
-			);
-
 		// Hash new password
 		const hashedNewPassword = await hashPassword(newPassword);
 
@@ -181,8 +160,12 @@ export const changePassword = async (req, res, next) => {
 	}
 };
 
-export const deleteAllUsers = async (req, res, next) => {
+export const deleteAll = async (req, res, next) => {
 	try {
+		// Get all users
+		const users = await User.find();
+		if (!users) return sendError(res, "Users not found", 404);
+
 		// Delete all users
 		await User.deleteMany();
 
@@ -193,7 +176,7 @@ export const deleteAllUsers = async (req, res, next) => {
 	}
 };
 
-export const deleteUserProfile = async (req, res, next) => {
+export const deleteById = async (req, res, next) => {
 	// Get user id from request params
 	const { id } = req.params;
 	// Get user id from request
